Manual/API

GeneralLedger.autoPullFiles — Method

autoPullFiles(path::AbstractString)

Call this function to read urls in local txt files line by line, and download each file

Example

julia> fl = autoPullFiles("c:mit.lailename.txt")

GeneralLedger.fixColNames — Method

fixColNames(str)

Call this function to update column names remove blank spaces, dollar sign, or Hash chars make all columns uppercase and replace hyphen with underscore

GeneralLedger.getWebLinks — Method

getWebLinks(url::AbstractString, fileTypes::AbstractVector, downloadPath:: AbstractString)

Call this function to crawl through a web page and download all file links first parameter is webpage url, next parameter is list of all file extensions user wish to download followed by output directory path

Example

julia> fl = getWebLinks("https://investor.apple.com/investor-relations/default.aspx#tabs_content--2021", ["pdf","csv","xlsx","xls"], "downloads/web")